Stops UserGetComments.gotData() crashing.
Files modified:
  joindinapp/Classes/UserGetComments.m

This method expects to receive an object comprising an NSDictionary of
NSDictionaries. Due to a change in the API (I'm trying to find out
what the change is) the remote server currently returns an error so that
the first dictionary comprises a single key/value pair being an error
message. The gotData() method then tries to call objectForKey() on the
message which results in a SIGABRT since it is an NSString and not an
NSDictionary.

This patch adds some robustness to the call testing the class types
and only processing the data if it is an NSDictionary object.